20180925-weisi-6.1.8(130): 1. 修复BleSensorProtocol在处理阵列传感器时有误的问题 2. 修复BleSensorProtocol的数据解析相关问题 3. 修复ScriptValueCorrector类correct方法强制转化结果错误的问题 4. 修复PhysicalSensor.addDynamicValue未将校正后数据代替原始数据

+++ weisi/src/main/java/com/cjq/lib/weisi/iot/corrector/ScriptValueCorrector.java
@@ -29,7 +29,7 @@ public class ScriptValueCorrector implements ValueCorrector {
     public double correct(double value) {
         mScript.getEngine().getContext().setAttribute("v", value, ScriptContext.ENGINE_SCOPE);
         try {
-            return (double) mScript.eval();
+            return ((Double) mScript.eval()).doubleValue();
         } catch (Exception e) {
             return value;
         }
